+++ react/components/EditorContainer/Sidebar/Content.tsx
@@ -45,10 +45,10 @@ const Content = (props: Props) => {
       if (path.current !== iframeRuntime.route.path) {
         setComponents(getInitialComponents(props))
         editor.setIsLoading(false)
+        path.current = iframeRuntime.route.path
       }
-      path.current = iframeRuntime.route.path
     },
-    [iframeRuntime]
+    [iframeRuntime.route.path]
   )
 
   if (editor.editTreePath === null) {