+++ react/components/admin/pages/Form/Form.tsx
@@ -1,8 +1,14 @@
 import React from 'react'
 import { defineMessages, FormattedMessage, injectIntl } from 'react-intl'
-import { Button, Checkbox, Input } from 'vtex.styleguide'
+import {
+  Button,
+  Checkbox,
+  EXPERIMENTAL_Select as Select,
+  Input,
+  Textarea,
+} from 'vtex.styleguide'
 
-import { RouteFormData } from 'pages'
+import { KeywordsFormData, RouteFormData } from 'pages'
 import FormFieldSeparator from '../../FormFieldSeparator'
 import SeparatorWithLine from '../SeparatorWithLine'
 
@@ -21,9 +27,12 @@ type TemplateSectionProps = Omit<
 interface CustomProps extends TemplateSectionProps {
   data: RouteFormData
   detailChangeHandlerGetter: (
-    detailName: keyof Route
-  ) => (event: React.ChangeEvent<HTMLInputElement>) => void
+    detailName: keyof RouteFormData
+  ) => (
+    event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
+  ) => void
   formErrors: FormErrors
+  isCustomPage: boolean
   isDeletable: boolean
   isInfoEditable: boolean
   isLoading: boolean
@@ -38,11 +47,16 @@ interface CustomProps extends TemplateSectionProps {
   ) => void
   onRemoveConditionalTemplate: (uniqueId: number) => void
   templates: Template[]
+  onChangeKeywords: (values: KeywordsFormData[]) => void
 }
 
 type Props = CustomProps & ReactIntl.InjectedIntlProps
 
 const messages = defineMessages({
+  createKeywordsMessage: {
+    defaultMessage: 'Create keyword "{keyword}"',
+    id: 'admin/pages.admin.pages.form.field.meta.keywords.create',
+  },
   detailsTitle: {
     defaultMessage: 'Page details',
     id: 'admin/pages.admin.pages.form.details.title',
@@ -59,6 +73,22 @@ const messages = defineMessages({
     defaultMessage: 'Title',
     id: 'admin/pages.admin.pages.form.field.title',
   },
+  noOptionsKeywordsMessage: {
+    defaultMessage: 'Type your keyword and press ENTER to add it.',
+    id: 'admin/pages.admin.pages.form.field.meta.description.no-options',
+  },
+  pathHint: {
+    defaultMessage: '/my-custom-page',
+    id: 'admin/pages.admin.pages.form.details.path-hint',
+  },
+  seoDescription: {
+    defaultMessage: 'Description',
+    id: 'admin/pages.admin.pages.form.field.meta.description',
+  },
+  seoKeywords: {
+    defaultMessage: 'Keywords',
+    id: 'admin/pages.admin.pages.form.field.meta.keywords',
+  },
 })
 
 const Form: React.FunctionComponent<Props> = ({
@@ -66,10 +96,12 @@ const Form: React.FunctionComponent<Props> = ({
   detailChangeHandlerGetter,
   formErrors,
   intl,
+  isCustomPage,
   isDeletable,
   isInfoEditable,
   isLoading,
   onAddConditionalTemplate,
+  onChangeKeywords,
   onChangeOperatorConditionalTemplate,
   onChangeStatementsConditionalTemplate,
   onChangeTemplateConditionalTemplate,
@@ -103,6 +135,7 @@ const Form: React.FunctionComponent<Props> = ({
         disabled={!isInfoEditable}
         label={intl.formatMessage(messages.fieldPath)}
         onChange={detailChangeHandlerGetter('path')}
+        placeholder={intl.formatMessage(messages.pathHint)}
         required
         value={path}
         errorMessage={
@@ -121,6 +154,33 @@ const Form: React.FunctionComponent<Props> = ({
         onChange={onLoginToggle}
         value="option-0"
       />
+      {isCustomPage && (
+        <>
+          <FormFieldSeparator />
+          <Textarea
+            disabled={!isInfoEditable}
+            label={intl.formatMessage(messages.seoDescription)}
+            onChange={detailChangeHandlerGetter('metaTagDescription')}
+            resize="vertical"
+            value={data.metaTagDescription}
+          />
+          <FormFieldSeparator />
+          <Select
+            creatable
+            disabled={!isInfoEditable}
+            formatCreateLabel={(keyword: string) =>
+              intl.formatMessage(messages.createKeywordsMessage, { keyword })
+            }
+            label={intl.formatMessage(messages.seoKeywords)}
+            noOptionsMessage={() =>
+              intl.formatMessage(messages.noOptionsKeywordsMessage)
+            }
+            onChange={onChangeKeywords}
+            placeholder=""
+            value={data.metaTagKeywords}
+          />
+        </>
+      )}
       <FormFieldSeparator />
       <SeparatorWithLine />
       <ConditionalTemplateSection