+++ react/typings/global.d.ts
@@ -55,6 +55,11 @@ declare global {
     statements: ConditionStatementArg[]
   }
 
+  interface PageMetaTags {
+    description?: string
+    keywords?: string[]
+  }
+
   interface Page {
     pageId?: string
     condition: ConditionArg
@@ -68,6 +73,7 @@ declare global {
     declarer: string | null
     domain: string
     interfaceId: string
+    metaTags?: PageMetaTags
     pages: Page[]
     path: string
     routeId: string