+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Image Uploader now has translated messages.
+- Loading issue when navigating.
 
 ## [3.11.0] - 2019-05-28